// front/src/model/GoodsCatalog.tsx
export interface GoodsCatalog{
    id_g: number;
    title: string;
    price: number;
    article: string;
    quantity: string;
    description: string;
    mark: number;
    composition: string;
    state: string;
    firstImage: string;
    typeId:number;
    shopId: number;
    createdAt: string;
    updatedAt: string;
}